header bug fixed and component directory arranged
