+++ client/src/Redux/action_creators/productActions.js
@@ -1,7 +1,7 @@
 import { fetchProducts, fetchAllProducts, fetchProductDetail } from '../../Api/product'
 import { addItemToCart, addItemToWishlist } from '../../Api/feature';
 
-//for sending list to store on first load
+//for sending product list to store
 export const putProducts = (productData) => {
     return {
         type: "PUT_PRODUCTS",
@@ -9,7 +9,7 @@ export const putProducts = (productData) => {
     }
 }
 
-// for getting list from api calls
+// for getting product list from api calls
 export const getProducts = (tag) => {
     return async dispatch => {
         try {
@@ -26,6 +26,7 @@ export const getProducts = (tag) => {
     }
 }
 
+//for sending product detail to store
 export const putProductDetail = (productDetailData) => {
     return {
         type: "PUT_PRODUCT_DETAIL",
@@ -33,6 +34,7 @@ export const putProductDetail = (productDetailData) => {
     }
 }
 
+// for getting product details from api calls
 export const getProductDetail = (id) => {
     return async dispatch => {
         try {
@@ -41,75 +43,5 @@ export const getProductDetail = (id) => {
         } catch (error) {
             console.log("error in getProductDetail", error)
         }
-    }
-}
-//need to add an api call and check if user is auth before calling api
-export const addComicToCart = (product, userInfo) => {
-    return async(dispatch) => {
-        try {
-            await addItemToCart(product, userInfo)
-            dispatch({type : "ADD_TO_CART", payload: product})
-        } catch (error) {
-            console.error("Error in productAction addComicToCart",error)
-        }
-    }
-}
-
-export const addComicToWishlist = (product, userInfo) => {
-    return async(dispatch) => {
-        try {
-            await addItemToWishlist(product, userInfo)
-            dispatch({type : "ADD_TO_WISHLIST", payload: product})
-        } catch (error) {
-            console.error("Error in productAction addItemToWishlist",error)
-        }
-    }
-}
-
-export const deleteItemFromCart = (id) => {
-    return async(dispatch) => {
-        try {
-            dispatch({type : "DELETE_ITEM_IN_CART", payload: id})
-        } catch (error) {
-            console.error("Error in productAction deleteItemFromCart",error)
-        }
-    }
-}
-
-export const handleCartQuantity = ( id, val) => {
-    return async(dispatch) => {
-        try {
-            if(val == '0'){
-                dispatch({type : "DELETE_ITEM_IN_CART", payload: id})
-            } else {
-                dispatch({
-                    type : "HANDLE_QUANTITY", 
-                    productId: id,
-                    value : val
-                })
-            }
-        } catch (error) {
-            console.error("Error in productAction handleCartQuantity",error)
-        }
-    }
-}
-
-export const clearItemFromCart = () => {
-    return async(dispatch) => {
-        try {
-            dispatch({type : "CLEAR_ITEM_IN_CART"})
-        } catch (error) {
-            console.error("Error in productAction handleCartQuantity",error)
-        }
-    }
-}
-
-export const deleteItemFromWishList = (id) => {
-    return async(dispatch) => {
-        try {
-            dispatch({type : "DELETE_ITEM_IN_WISHLIST", payload: id})
-        } catch (error) {
-            console.error("Error in productAction deleteItemFromWishList",error)
-        } 
     }
 }
